% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimatePercentageErrors.R
\name{estimatePercentageErrors}
\alias{estimatePercentageErrors}
\title{Estimate percentage errors}
\usage{
estimatePercentageErrors(x, conf.level = 0.95, coord.flip = FALSE,
  digits = 5)
}
\arguments{
\item{x:}{The data frame containing colums category, cases and total.}

\item{conf.level:}{Confidence level for the returned confidence interval, default 0.95.}

\item{coord.flip:}{Cartesian coordinates with x and y flipped, default FALSE.}

\item{digits:}{Integer indicating the number of decimal places to be used, default 5.}
}
\value{
\code{estimatePercentageErrors} returns a a barplot and table of results.
}
\description{
Functions to estimate percentage errors
}
\examples{

estimatePercentageErrors(data20)
estimatePercentageErrors(data20, conf.level = 0.8)
estimatePercentageErrors(mydata, conf.level = 0.8, coord.flip = TRUE)
estimatePercentageErrors(mydata, conf.level = 0.8, coord.flip = TRUE, digits = 2)

}
\author{
Andrey Davydenko, Maxim and Sai Van Cuong from Volgograd State Technical University
}
