% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcbars-package.R
\docType{package}
\name{pcbars-package}
\alias{pcbars-package}
\alias{pcbars}
\title{Estimate percentage errors}
\source{
\url{https://github.com/svcuonghvktqs/pcbars}.
}
\description{
Contains tools to measure and visualise the uncertainty around the percentages used to represent relative frequencies. These are useful when performing descriptive analysis of categorical variables.
}
\details{
{\tabular{ll}{ Package: \tab pcbars\cr Type: \tab Package\cr License: \tab GPL3\cr}}
}
\author{
Andrey Davydenko, Maxim Shcherbakov and Sai Van Cuong from Volgograd State Technical University
}
\keyword{package}
